Separate expression compilation from the Compiler
Take the functions that compile expressions out of the Compiler type so
that they are standalone.  This clearly separates side-effects from the
compilation logic and is how it should have been done in the first
place.

This is necessary to support invoking expression compilation separately
from the upcoming command and function argument parsing hooks.
