+++ core/src/compiler/mod.rs
@@ -22,6 +22,8 @@ use crate::syms::{Symbol, SymbolKey, Symbols};
 use std::collections::HashMap;
 use std::fmt;
 
+mod exprs;
+
 /// Compilation errors.
 #[derive(Debug, thiserror::Error)]
 #[error("{}:{}: {}", pos.line, pos.col, message)]
@@ -332,30 +334,15 @@ impl Compiler {
         args: Vec<Expr>,
         name_pos: LineCol,
     ) -> Result<()> {
-        if exp_nargs != args.len() {
-            return Err(Error::new(
-                name_pos,
-                format!("Cannot index array with {} subscripts; need {}", args.len(), exp_nargs),
-            ));
-        }
-
-        for arg in args.into_iter().rev() {
-            let arg_pos = arg.start_pos();
-            match self.compile_expr(arg, false)? {
-                ExprType::Integer => (),
-                ExprType::Double => {
-                    self.emit(Instruction::DoubleToInteger);
-                }
-                itype => {
-                    return Err(Error::new(
-                        arg_pos,
-                        format!("Array index must be INTEGER, not {}", itype),
-                    ));
-                }
+        let mut instrs = vec![];
+        match exprs::compile_array_indices(&mut instrs, &self.symtable, exp_nargs, args, name_pos) {
+            Ok(result) => {
+                self.next_pc += instrs.len();
+                self.instrs.append(&mut instrs);
+                Ok(result)
             }
+            Err(e) => Err(e),
         }
-
-        Ok(())
     }
 
     /// Compiles an assignment to an array position.
@@ -719,415 +706,6 @@ impl Compiler {
         Ok(())
     }
 
-    /// Compiles a logical or bitwise unary operator and appends its instructions to the compilation
-    /// context.
-    fn compile_not_op(&mut self, span: UnaryOpSpan) -> Result<ExprType> {
-        let expr_type = self.compile_expr(span.expr, false)?;
-        match expr_type {
-            ExprType::Boolean => {
-                self.emit(Instruction::LogicalNot(span.pos));
-                Ok(ExprType::Boolean)
-            }
-            ExprType::Integer => {
-                self.emit(Instruction::BitwiseNot(span.pos));
-                Ok(ExprType::Integer)
-            }
-            _ => Err(Error::new(span.pos, format!("Cannot apply NOT to {}", expr_type))),
-        }
-    }
-
-    /// Compiles a negate operator and appends its instructions to the compilation context.
-    fn compile_neg_op(&mut self, span: UnaryOpSpan) -> Result<ExprType> {
-        let expr_type = self.compile_expr(span.expr, false)?;
-        match expr_type {
-            ExprType::Double => {
-                self.emit(Instruction::NegateDouble(span.pos));
-                Ok(ExprType::Double)
-            }
-            ExprType::Integer => {
-                self.emit(Instruction::NegateInteger(span.pos));
-                Ok(ExprType::Integer)
-            }
-            _ => Err(Error::new(span.pos, format!("Cannot negate {}", expr_type))),
-        }
-    }
-
-    /// Compiles a logical binary operator and appends its instructions to the compilation context.
-    fn compile_logical_binary_op<F1: Fn(LineCol) -> Instruction, F2: Fn(LineCol) -> Instruction>(
-        &mut self,
-        logical_make_inst: F1,
-        bitwise_make_inst: F2,
-        span: BinaryOpSpan,
-        op_name: &str,
-    ) -> Result<ExprType> {
-        let lhs_type = self.compile_expr(span.lhs, false)?;
-        let rhs_type = self.compile_expr(span.rhs, false)?;
-        match (lhs_type, rhs_type) {
-            (ExprType::Boolean, ExprType::Boolean) => {
-                self.emit(logical_make_inst(span.pos));
-                Ok(ExprType::Boolean)
-            }
-            (ExprType::Integer, ExprType::Integer) => {
-                self.emit(bitwise_make_inst(span.pos));
-                Ok(ExprType::Integer)
-            }
-            (_, _) => Err(Error::new(
-                span.pos,
-                format!("Cannot {} {} and {}", op_name, lhs_type, rhs_type),
-            )),
-        }
-    }
-
-    /// Compiles an equality binary operator and appends its instructions to the compilation
-    /// context.
-    fn compile_equality_binary_op<
-        F1: Fn(LineCol) -> Instruction,
-        F2: Fn(LineCol) -> Instruction,
-        F3: Fn(LineCol) -> Instruction,
-        F4: Fn(LineCol) -> Instruction,
-    >(
-        &mut self,
-        boolean_make_inst: F1,
-        double_make_inst: F2,
-        integer_make_inst: F3,
-        text_make_inst: F4,
-        span: BinaryOpSpan,
-        op_name: &str,
-    ) -> Result<ExprType> {
-        let lhs_type = self.compile_expr(span.lhs, false)?;
-        let pc = self.next_pc;
-        self.emit(Instruction::Nop);
-
-        let mut keep_nop = false;
-        let rhs_type = self.compile_expr(span.rhs, false)?;
-        let result = match (lhs_type, rhs_type) {
-            (lhs_type, rhs_type) if lhs_type == rhs_type => lhs_type,
-
-            (ExprType::Double, ExprType::Integer) => {
-                self.emit(Instruction::IntegerToDouble);
-                ExprType::Double
-            }
-
-            (ExprType::Integer, ExprType::Double) => {
-                self.instrs[pc] = Instruction::IntegerToDouble;
-                keep_nop = true;
-                ExprType::Double
-            }
-
-            (_, _) => {
-                return Err(Error::new(
-                    span.pos,
-                    format!("Cannot compare {} and {} with {}", lhs_type, rhs_type, op_name),
-                ));
-            }
-        };
-
-        if !keep_nop {
-            self.instrs.remove(pc);
-            self.next_pc -= 1;
-        }
-
-        match result {
-            ExprType::Boolean => self.emit(boolean_make_inst(span.pos)),
-            ExprType::Double => self.emit(double_make_inst(span.pos)),
-            ExprType::Integer => self.emit(integer_make_inst(span.pos)),
-            ExprType::Text => self.emit(text_make_inst(span.pos)),
-        };
-
-        Ok(ExprType::Boolean)
-    }
-
-    /// Compiles a relational binary operator and appends its instructions to the compilation
-    /// context.
-    fn compile_relational_binary_op<
-        F1: Fn(LineCol) -> Instruction,
-        F2: Fn(LineCol) -> Instruction,
-        F3: Fn(LineCol) -> Instruction,
-    >(
-        &mut self,
-        double_make_inst: F1,
-        integer_make_inst: F2,
-        text_make_inst: F3,
-        span: BinaryOpSpan,
-        op_name: &str,
-    ) -> Result<ExprType> {
-        let lhs_type = self.compile_expr(span.lhs, false)?;
-        let pc = self.next_pc;
-        self.emit(Instruction::Nop);
-
-        let mut keep_nop = false;
-        let rhs_type = self.compile_expr(span.rhs, false)?;
-        let result = match (lhs_type, rhs_type) {
-            // Boolean is explicitly excluded here.
-            (ExprType::Double, ExprType::Double) => ExprType::Double,
-            (ExprType::Integer, ExprType::Integer) => ExprType::Integer,
-            (ExprType::Text, ExprType::Text) => ExprType::Text,
-
-            (ExprType::Double, ExprType::Integer) => {
-                self.emit(Instruction::IntegerToDouble);
-                ExprType::Double
-            }
-
-            (ExprType::Integer, ExprType::Double) => {
-                self.instrs[pc] = Instruction::IntegerToDouble;
-                keep_nop = true;
-                ExprType::Double
-            }
-
-            (_, _) => {
-                return Err(Error::new(
-                    span.pos,
-                    format!("Cannot compare {} and {} with {}", lhs_type, rhs_type, op_name),
-                ));
-            }
-        };
-
-        if !keep_nop {
-            self.instrs.remove(pc);
-            self.next_pc -= 1;
-        }
-
-        match result {
-            ExprType::Boolean => unreachable!("Filtered out above"),
-            ExprType::Double => self.emit(double_make_inst(span.pos)),
-            ExprType::Integer => self.emit(integer_make_inst(span.pos)),
-            ExprType::Text => self.emit(text_make_inst(span.pos)),
-        };
-
-        Ok(ExprType::Boolean)
-    }
-
-    /// Compiles a binary shift operator and appends its instructions to the compilation context.
-    fn compile_shift_binary_op<F: Fn(LineCol) -> Instruction>(
-        &mut self,
-        make_inst: F,
-        span: BinaryOpSpan,
-        op_name: &str,
-    ) -> Result<ExprType> {
-        let lhs_type = self.compile_expr(span.lhs, false)?;
-        match lhs_type {
-            ExprType::Integer => (),
-            _ => {
-                return Err(Error::new(
-                    span.pos,
-                    format!("Cannot apply {} to {}", op_name, lhs_type),
-                ));
-            }
-        };
-
-        let rhs_type = self.compile_expr(span.rhs, false)?;
-        match rhs_type {
-            ExprType::Integer => (),
-            _ => {
-                return Err(Error::new(
-                    span.pos,
-                    format!(
-                        "Number of bits to {} must be an {}, not a {}",
-                        op_name,
-                        ExprType::Integer,
-                        rhs_type
-                    ),
-                ));
-            }
-        };
-
-        self.emit(make_inst(span.pos));
-
-        Ok(ExprType::Integer)
-    }
-
-    /// Compiles the evaluation of an expression, appends its instructions to the
-    /// Compiles a binary operator and appends its instructions to the compilation context.
-    fn compile_arithmetic_binary_op<
-        F1: Fn(LineCol) -> Instruction,
-        F2: Fn(LineCol) -> Instruction,
-    >(
-        &mut self,
-        double_make_inst: F1,
-        integer_make_inst: F2,
-        span: BinaryOpSpan,
-        op_name: &str,
-    ) -> Result<ExprType> {
-        let lhs_type = self.compile_expr(span.lhs, false)?;
-        let pc = self.next_pc;
-        self.emit(Instruction::Nop);
-
-        let mut keep_nop = false;
-        let rhs_type = self.compile_expr(span.rhs, false)?;
-        let result = match (lhs_type, rhs_type) {
-            (ExprType::Double, ExprType::Double) => ExprType::Double,
-            (ExprType::Integer, ExprType::Integer) => ExprType::Integer,
-            (ExprType::Text, ExprType::Text) if op_name == "+" => ExprType::Text,
-
-            (ExprType::Double, ExprType::Integer) => {
-                self.emit(Instruction::IntegerToDouble);
-                ExprType::Double
-            }
-
-            (ExprType::Integer, ExprType::Double) => {
-                self.instrs[pc] = Instruction::IntegerToDouble;
-                keep_nop = true;
-                ExprType::Double
-            }
-
-            (_, _) => {
-                return Err(Error::new(
-                    span.pos,
-                    format!("Cannot {} {} and {}", op_name, lhs_type, rhs_type),
-                ));
-            }
-        };
-
-        if !keep_nop {
-            self.instrs.remove(pc);
-            self.next_pc -= 1;
-        }
-
-        match result {
-            ExprType::Boolean => unreachable!("Filtered out above"),
-            ExprType::Double => self.emit(double_make_inst(span.pos)),
-            ExprType::Integer => self.emit(integer_make_inst(span.pos)),
-            ExprType::Text => {
-                debug_assert_eq!("+", op_name, "Filtered out above");
-                self.emit(Instruction::ConcatStrings(span.pos))
-            }
-        };
-
-        Ok(result)
-    }
-
-    /// Compiles the load of a symbol in the context of an expression.
-    fn compile_expr_symbol(&mut self, span: SymbolSpan) -> Result<ExprType> {
-        let key = SymbolKey::from(span.vref.name());
-        let (instr, vtype) = match self.symtable.get(&key) {
-            None => {
-                return Err(Error::new(
-                    span.pos,
-                    format!("Undefined variable {}", span.vref.name()),
-                ))
-            }
-
-            Some(SymbolPrototype::Variable(vtype)) => (Instruction::Load(key, span.pos), vtype),
-
-            Some(SymbolPrototype::Callable(ctype, is_argless)) => {
-                if !is_argless {
-                    return Err(Error::new(
-                        span.pos,
-                        format!("In call to {}: function requires arguments", span.vref.name()),
-                    ));
-                }
-
-                match ctype.as_expr_type() {
-                    Some(etype) => {
-                        (Instruction::FunctionCall(key, ctype.into(), span.pos, 0), etype)
-                    }
-                    None => {
-                        return Err(Error::new(
-                            span.pos,
-                            format!("{} is not an array nor a function", span.vref.name()),
-                        ));
-                    }
-                }
-            }
-
-            Some(_) => {
-                return Err(Error::new(span.pos, format!("{} is not a variable", span.vref.name())))
-            }
-        };
-        if !span.vref.accepts(vtype.into()) {
-            return Err(Error::new(
-                span.pos,
-                format!("Incompatible type annotation in {} reference", span.vref),
-            ));
-        }
-        self.emit(instr);
-        Ok(vtype)
-    }
-
-    /// Compiles the load of a symbol in the context of a command argument.
-    fn compile_expr_symbol_ref(&mut self, span: SymbolSpan) -> Result<ExprType> {
-        let key = SymbolKey::from(span.vref.name());
-        let result = match self.symtable.get(&key) {
-            None => {
-                let vtype = if span.vref.ref_type() == VarType::Auto {
-                    ExprType::Integer
-                } else {
-                    span.vref.ref_type().into()
-                };
-
-                if !span.vref.accepts(vtype.into()) {
-                    return Err(Error::new(
-                        span.pos,
-                        format!("Incompatible type annotation in {} reference", span.vref),
-                    ));
-                }
-
-                self.symtable.insert(key, SymbolPrototype::Variable(vtype));
-                vtype
-            }
-
-            Some(SymbolPrototype::Array(vtype, _)) | Some(SymbolPrototype::Variable(vtype)) => {
-                if !span.vref.accepts(vtype.into()) {
-                    return Err(Error::new(
-                        span.pos,
-                        format!("Incompatible type annotation in {} reference", span.vref),
-                    ));
-                }
-
-                vtype
-            }
-
-            Some(SymbolPrototype::Callable(ctype, is_argless)) => match ctype.as_expr_type() {
-                Some(etype) => {
-                    if !span.vref.accepts(etype.into()) {
-                        return Err(Error::new(
-                            span.pos,
-                            format!("Incompatible type annotation in {} reference", span.vref),
-                        ));
-                    }
-
-                    if !is_argless {
-                        return Err(Error::new(
-                            span.pos,
-                            format!("In call to {}: function requires arguments", span.vref.name()),
-                        ));
-                    }
-
-                    etype
-                }
-                None => {
-                    return Err(Error::new(
-                        span.pos,
-                        format!("{} is not an array nor a function", span.vref.name()),
-                    ));
-                }
-            },
-        };
-        self.emit(Instruction::LoadRef(span.vref, span.pos));
-        Ok(result)
-    }
-
-    /// Compiles an array access.
-    fn compile_array_ref(
-        &mut self,
-        span: FunctionCallSpan,
-        key: SymbolKey,
-        vtype: ExprType,
-        dimensions: usize,
-    ) -> Result<ExprType> {
-        let nargs = span.args.len();
-        self.compile_array_indices(dimensions, span.args, span.pos)?;
-
-        if !span.fref.accepts(vtype.into()) {
-            return Err(Error::new(
-                span.pos,
-                format!("Incompatible type annotation in {} reference", span.fref),
-            ));
-        }
-        self.emit(Instruction::ArrayLoad(key, span.pos, nargs));
-        Ok(vtype)
-    }
-
     /// Compiles the evaluation of an expression, appends its instructions to the
     /// compilation context, and returns the type of the compiled expression.
     ///
@@ -1136,243 +714,17 @@ impl Compiler {
     /// references unevaluated because we don't know if the function wants to take
     /// the reference or the value.
     fn compile_expr(&mut self, expr: Expr, allow_varrefs: bool) -> Result<ExprType> {
-        match expr {
-            Expr::Boolean(span) => {
-                self.emit(Instruction::Push(Value::Boolean(span.value), span.pos));
-                Ok(ExprType::Boolean)
-            }
-
-            Expr::Double(span) => {
-                self.emit(Instruction::Push(Value::Double(span.value), span.pos));
-                Ok(ExprType::Double)
-            }
-
-            Expr::Integer(span) => {
-                self.emit(Instruction::Push(Value::Integer(span.value), span.pos));
-                Ok(ExprType::Integer)
-            }
-
-            Expr::Text(span) => {
-                self.emit(Instruction::Push(Value::Text(span.value), span.pos));
-                Ok(ExprType::Text)
-            }
-
-            Expr::Symbol(span) => {
-                if allow_varrefs {
-                    self.compile_expr_symbol_ref(span)
-                } else {
-                    self.compile_expr_symbol(span)
-                }
-            }
-
-            Expr::And(span) => self.compile_logical_binary_op(
-                Instruction::LogicalAnd,
-                Instruction::BitwiseAnd,
-                *span,
-                "AND",
-            ),
-
-            Expr::Or(span) => self.compile_logical_binary_op(
-                Instruction::LogicalOr,
-                Instruction::BitwiseOr,
-                *span,
-                "OR",
-            ),
-
-            Expr::Xor(span) => self.compile_logical_binary_op(
-                Instruction::LogicalXor,
-                Instruction::BitwiseXor,
-                *span,
-                "XOR",
-            ),
-
-            Expr::Not(span) => self.compile_not_op(*span),
-
-            Expr::ShiftLeft(span) => {
-                let result = self.compile_shift_binary_op(Instruction::ShiftLeft, *span, "<<")?;
-                debug_assert_eq!(ExprType::Integer, result);
+        let result = if allow_varrefs {
+            exprs::compile_expr_in_command(&mut self.instrs, &mut self.symtable, expr)
+        } else {
+            exprs::compile_expr(&mut self.instrs, &self.symtable, expr, false)
+        };
+        match result {
+            Ok(result) => {
+                self.next_pc = self.instrs.len();
                 Ok(result)
             }
-
-            Expr::ShiftRight(span) => {
-                let result = self.compile_shift_binary_op(Instruction::ShiftRight, *span, ">>")?;
-                debug_assert_eq!(ExprType::Integer, result);
-                Ok(result)
-            }
-
-            Expr::Equal(span) => {
-                let result = self.compile_equality_binary_op(
-                    Instruction::EqualBooleans,
-                    Instruction::EqualDoubles,
-                    Instruction::EqualIntegers,
-                    Instruction::EqualStrings,
-                    *span,
-                    "=",
-                )?;
-                debug_assert_eq!(ExprType::Boolean, result);
-                Ok(result)
-            }
-
-            Expr::NotEqual(span) => {
-                let result = self.compile_equality_binary_op(
-                    Instruction::NotEqualBooleans,
-                    Instruction::NotEqualDoubles,
-                    Instruction::NotEqualIntegers,
-                    Instruction::NotEqualStrings,
-                    *span,
-                    "<>",
-                )?;
-                debug_assert_eq!(ExprType::Boolean, result);
-                Ok(result)
-            }
-
-            Expr::Less(span) => {
-                let result = self.compile_relational_binary_op(
-                    Instruction::LessDoubles,
-                    Instruction::LessIntegers,
-                    Instruction::LessStrings,
-                    *span,
-                    "<",
-                )?;
-                debug_assert_eq!(ExprType::Boolean, result);
-                Ok(result)
-            }
-
-            Expr::LessEqual(span) => {
-                let result = self.compile_relational_binary_op(
-                    Instruction::LessEqualDoubles,
-                    Instruction::LessEqualIntegers,
-                    Instruction::LessEqualStrings,
-                    *span,
-                    "<=",
-                )?;
-                debug_assert_eq!(ExprType::Boolean, result);
-                Ok(result)
-            }
-
-            Expr::Greater(span) => {
-                let result = self.compile_relational_binary_op(
-                    Instruction::GreaterDoubles,
-                    Instruction::GreaterIntegers,
-                    Instruction::GreaterStrings,
-                    *span,
-                    ">",
-                )?;
-                debug_assert_eq!(ExprType::Boolean, result);
-                Ok(result)
-            }
-
-            Expr::GreaterEqual(span) => {
-                let result = self.compile_relational_binary_op(
-                    Instruction::GreaterEqualDoubles,
-                    Instruction::GreaterEqualIntegers,
-                    Instruction::GreaterEqualStrings,
-                    *span,
-                    ">=",
-                )?;
-                debug_assert_eq!(ExprType::Boolean, result);
-                Ok(result)
-            }
-
-            Expr::Add(span) => Ok(self.compile_arithmetic_binary_op(
-                Instruction::AddDoubles,
-                Instruction::AddIntegers,
-                *span,
-                "+",
-            )?),
-
-            Expr::Subtract(span) => Ok(self.compile_arithmetic_binary_op(
-                Instruction::SubtractDoubles,
-                Instruction::SubtractIntegers,
-                *span,
-                "-",
-            )?),
-
-            Expr::Multiply(span) => Ok(self.compile_arithmetic_binary_op(
-                Instruction::MultiplyDoubles,
-                Instruction::MultiplyIntegers,
-                *span,
-                "*",
-            )?),
-
-            Expr::Divide(span) => Ok(self.compile_arithmetic_binary_op(
-                Instruction::DivideDoubles,
-                Instruction::DivideIntegers,
-                *span,
-                "/",
-            )?),
-
-            Expr::Modulo(span) => Ok(self.compile_arithmetic_binary_op(
-                Instruction::ModuloDoubles,
-                Instruction::ModuloIntegers,
-                *span,
-                "MOD",
-            )?),
-
-            Expr::Power(span) => Ok(self.compile_arithmetic_binary_op(
-                Instruction::PowerDoubles,
-                Instruction::PowerIntegers,
-                *span,
-                "^",
-            )?),
-
-            Expr::Negate(span) => Ok(self.compile_neg_op(*span)?),
-
-            Expr::Call(span) => {
-                let key = SymbolKey::from(span.fref.name());
-                match self.symtable.get(&key) {
-                    Some(SymbolPrototype::Array(vtype, dims)) => {
-                        self.compile_array_ref(span, key, vtype, dims)
-                    }
-
-                    Some(SymbolPrototype::Callable(ctype, is_argless)) => {
-                        let nargs = span.args.len();
-                        for arg in span.args.into_iter().rev() {
-                            self.compile_expr(arg, true)?;
-                        }
-
-                        if !span.fref.accepts((ctype).into()) {
-                            return Err(Error::new(
-                                span.pos,
-                                format!("Incompatible type annotation in {} reference", span.fref),
-                            ));
-                        }
-
-                        let vtype = match ctype.as_expr_type() {
-                            Some(vtype) => vtype,
-                            None => {
-                                return Err(Error::new(
-                                    span.pos,
-                                    format!("{} is not an array nor a function", span.fref.name()),
-                                ));
-                            }
-                        };
-
-                        if is_argless {
-                            return Err(Error::new(
-                                span.pos,
-                                format!(
-                                    "In call to {}: expected no arguments nor parenthesis",
-                                    span.fref.name()
-                                ),
-                            ));
-                        }
-
-                        self.emit(Instruction::FunctionCall(key, ctype.into(), span.pos, nargs));
-                        Ok(vtype)
-                    }
-
-                    Some(SymbolPrototype::Variable(_)) => Err(Error::new(
-                        span.pos,
-                        format!("{} is not an array nor a function", span.fref.name()),
-                    )),
-
-                    None => Err(Error::new(
-                        span.pos,
-                        format!("Unknown function or array {}", span.fref.name()),
-                    )),
-                }
-            }
+            Err(e) => Err(e),
         }
     }
